requirements.txt
aniso8601==8.0.0
APScheduler==3.6.3
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
click==7.1.1
colorama==0.4.3
Flask==1.1.1
Flask-Admin==1.5.6
Flask-API==2.0
Flask-MySQL==1.5.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
gitdb==4.0.4
GitPython==3.1.1
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
mysqlclient==1.4.6
numpy==1.18.2
packaging==20.3
pandas==1.0.3
Pillow==7.1.0
pluggy==0.13.1
py==1.8.1
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
schedule==0.6.0
six==1.14.0
smmap==3.0.2
SQLAlchemy==1.3.15
svgwrite==1.4
Tree==0.2.4
typed-ast==1.4.1
tzlocal==2.0.0
wcwidth==0.1.9
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
zipp==3.1.0
